誤記修正

+++ app/src/main/java/com/daigorian/epcltvapp/epgstationv2caller/Rules.kt
@@ -2,5 +2,5 @@ package com.daigorian.epcltvapp.epgstationv2caller
 //ルール情報
 data class Rules(
     val rules : List<Rule>?,
-    val total : Integer? //ルール総件数
+    val total : Long? //ルール総件数
 )
